global API change to use "int" instead of "unsigned int"
This is an API change that should not break the ABI.

The rationale is that "a number of ..." is expressed as an int in almost
all cases.
It has the nice side-effect of being shorter to write (and read).
It should make the API simpler and the code using the rtfilter API
easier to write and read.

API variables changed:
* ns, r (number of samples)
* nchann (number of channels)
* numlen, denum_len (number of coefficients)
* num_poles (number of poles)
* hlen (fir half-length)

+++ doc/examples/butterworth.c
@@ -35,11 +35,11 @@
 
 /* Create a ramp signal in each channel */
 static
-void create_input(float* data, unsigned int ns, unsigned int cs)
+void create_input(float* data, int ns, int cs)
 {
-	unsigned int j, i, ramplen;
+	int j, i, ramplen;
 
-	ramplen = (unsigned int)(RAMPDUR*FS);
+	ramplen = RAMPDUR * FS;
 
 	for (j = 0; j < NCH; j++) {
 		for (i = 0; i < ns; i++) {
@@ -56,7 +56,7 @@ int main(int argc, char* argv[])
 
 	float input[CHUNKNS * NCH], output[CHUNKNS * NCH];
 	double normfc;
-	unsigned int i;
+	int i;
 	hfilter filt = NULL;
 	FILE *fin = NULL, *fout = NULL;
 	size_t input_ns, output_ns;
